Impersontation steps refactor
