+++ Projects/RMAutomationUIFramework/src/test/java/org/fundacionjala/automation/scenario/steps/admin/location/LocationGivenSteps.java
@@ -4,6 +4,7 @@ import org.fundacionjala.automation.framework.pages.admin.login.LoginPage;
 import org.fundacionjala.automation.framework.utils.api.managers.LocationAPIManager;
 import org.fundacionjala.automation.framework.utils.api.objects.admin.Location;
 import org.fundacionjala.automation.framework.utils.common.BrowserManager;
+import org.fundacionjala.automation.framework.utils.common.PropertiesReader;
 
 import cucumber.api.java.en.Given;
 
@@ -23,6 +24,6 @@ public class LocationGivenSteps {
 	
 	@Given("^I have a location added with name: \"([^\"]*)\", display name \"([^\"]*)\" and description \"([^\"]*)\"$")
 	public void i_have_a_location_added_with_name_display_name_and_description(String name, String displayName, String description) throws Throwable {
-		LocationAPIManager.postRequest("http://172.20.208.84:4040/locations", new Location(name, displayName, "", "", description));
+		LocationAPIManager.postRequest(PropertiesReader.getServiceURL() +"/locations", new Location(name, displayName, "", "", description));
 	}
 }
